Initial Fence Sync Object support
Allows creating and managing binary state sync
objects.  Currently they aren't useful because
there is not yet a way to wait for them or
query their state.  X fence objects are owned
by a screen.  As Aaron Plattner pointed out,
screens are identified by a drawable in X
protocol, so XSyncCreateFence() takes a drawable
to identify which screen to create the fence on.

Signed-off-by: James Jones <jajones@nvidia.com>
Reviewed-by: Aaron Plattner <aplattner@nvidia.com>
Acked-by: Alan Coopersmith <alan.coopersmith@oracle.com>