Correct XShm return values.
XShmAttach, XShmDetach, XShmPutImage do not return a Status but 0 or 1.
Though the man section for XShmAttach says "if all goes well, you will get a
non-zero status, back" this is counter to the usage of Status in Xlib
itself where 0 means Success and no-zero specifies the specific error.

XShmPixmapFormat does not return a Status but the pixmap format or 0 on
failure.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Adam Jackson <ajax@redhat.com>